# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ScorpStudy by Bishal API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LimitError {
  error: string;
  message: string;
}

export interface AiResponse {
  response: string;
  fromCache: boolean;
  provider?: string;
}

export interface ChatMessage {
  role: string;
  content: string;
}

export interface ChatInput {
  message: string;
  subject?: string;
  conversationHistory?: ChatMessage[];
  userId: string;
}

export interface QuizInput {
  topic?: string;
  notes?: string;
  numQuestions?: number;
  difficulty?: string;
  questionType?: string;
  userId: string;
}

export interface QuizQuestion {
  question: string;
  options?: string[];
  correctAnswer: string;
  explanation?: string;
  type: string;
}

export interface QuizResponse {
  questions: QuizQuestion[];
  topic?: string;
}

export interface FlashcardInput {
  topic?: string;
  notes?: string;
  numCards?: number;
  userId: string;
}

export interface Flashcard {
  front: string;
  back: string;
}

export interface FlashcardResponse {
  cards: Flashcard[];
  topic?: string;
}

export type SummarizeInputLength = typeof SummarizeInputLength[keyof typeof SummarizeInputLength];


export const SummarizeInputLength = {
  brief: 'brief',
  standard: 'standard',
  detailed: 'detailed',
} as const;

export interface SummarizeInput {
  text?: string;
  url?: string;
  length?: SummarizeInputLength;
  userId: string;
}

export interface VocabWord {
  word: string;
  definition: string;
}

export interface SummaryResponse {
  summary: string;
  keyPoints: string[];
  examQuestions: string[];
  vocabulary: VocabWord[];
}

export interface TranslateInput {
  text: string;
  targetLanguage: string;
  sourceLanguage?: string;
  userId: string;
}

export interface TranslateResponse {
  translation: string;
  detectedLanguage?: string;
  provider?: string;
}

export interface MindmapInput {
  topic?: string;
  notes?: string;
  userId: string;
}

export interface MindmapNode {
  label: string;
  children: string[];
}

export interface MindmapResponse {
  center: string;
  branches: MindmapNode[];
}

export interface EnhancePromptInput {
  prompt: string;
  userId: string;
}

export interface PromptEnhancement {
  enhancedPrompt: string;
}

export interface FormulaInput {
  question: string;
  userId: string;
}

export interface NoteAiInput {
  content: string;
  userId: string;
}

export interface Note {
  id: number;
  userId: string;
  title: string;
  content: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  title: string;
  content?: string;
  tags?: string[];
  userId: string;
}

export interface NoteUpdate {
  title?: string;
  content?: string;
  tags?: string[];
}

export interface DailyUsage {
  userId: string;
  aiQueries: number;
  imagesGenerated: number;
  aiQueriesLimit: number;
  imagesLimit: number;
  date: string;
}

export interface QuizResult {
  id: number;
  userId: string;
  topic?: string;
  score: number;
  total: number;
  percentage: number;
  difficulty?: string;
  questionType?: string;
  createdAt: string;
}

export interface QuizResultInput {
  topic?: string;
  score: number;
  total: number;
  percentage: number;
  difficulty?: string;
  questionType?: string;
  userId: string;
}

export interface FlashcardSet {
  id: number;
  userId: string;
  topic: string;
  cards: Flashcard[];
  createdAt: string;
}

export interface FlashcardSetInput {
  topic: string;
  cards: Flashcard[];
  userId: string;
}

export interface Summary {
  id: number;
  userId: string;
  sourceType?: string;
  originalText?: string;
  summary: string;
  keyPoints: string[];
  examQuestions: string[];
  createdAt: string;
}

export interface SummaryInput {
  sourceType?: string;
  originalText?: string;
  summary: string;
  keyPoints?: string[];
  examQuestions?: string[];
  userId: string;
}

export interface GeneratedImage {
  id: number;
  userId: string;
  prompt: string;
  /** @nullable */
  enhancedPrompt?: string | null;
  imageUrl: string;
  style?: string;
  createdAt: string;
}

export interface GeneratedImageInput {
  prompt: string;
  /** @nullable */
  enhancedPrompt?: string | null;
  imageUrl: string;
  style?: string;
  userId: string;
}

export interface Mindmap {
  id: number;
  userId: string;
  topic: string;
  mapData: MindmapResponse;
  createdAt: string;
}

export interface MindmapSaveInput {
  topic: string;
  mapData: MindmapResponse;
  userId: string;
}

export interface Translation {
  id: number;
  userId: string;
  originalText: string;
  translatedText: string;
  sourceLanguage?: string;
  targetLanguage: string;
  createdAt: string;
}

export interface TranslationInput {
  originalText: string;
  translatedText: string;
  sourceLanguage?: string;
  targetLanguage: string;
  userId: string;
}

export interface HistoryItem {
  id: string;
  type: string;
  title: string;
  description?: string;
  createdAt: string;
}

export interface DayCount {
  date: string;
  count: number;
}

export interface FeatureCount {
  feature: string;
  count: number;
}

export interface Analytics {
  totalQueriesThisWeek: number;
  totalQueriesSavedByCache: number;
  studyStreakDays: number;
  mostUsedFeature?: string;
  queriesByDay: DayCount[];
  usageByFeature: FeatureCount[];
}

export interface HistoryResponse {
  items: HistoryItem[];
  analytics: Analytics;
}

export type ListNotesParams = {
search?: string;
sort?: ListNotesSort;
};

export type ListNotesSort = typeof ListNotesSort[keyof typeof ListNotesSort];


export const ListNotesSort = {
  newest: 'newest',
  oldest: 'oldest',
  title: 'title',
} as const;