# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScorpStudy by Bishal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/enhance-prompt:
    post:
      operationId: enhancePrompt
      tags: [ai]
      summary: Enhance an image generation prompt with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnhancePromptInput"
      responses:
        "200":
          description: Enhanced prompt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromptEnhancement"
        "429":
          description: Daily limit reached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LimitError"
...
components:
  schemas:
...
    LimitError:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    EnhancePromptInput:
      type: object
      properties:
        prompt:
          type: string
        userId:
          type: string
      required:
        - prompt
        - userId
    PromptEnhancement:
      type: object
      properties:
        enhancedPrompt:
          type: string
      required:
        - enhancedPrompt
...