# lib/api-zod/src/generated/types/limitError.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * ScorpStudy by Bishal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LimitError {
  error: string;
  message: string;
}